+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed accessing settings for PillowEncoder.
+
 ## [0.14.0] - 2023-11-29
 
 ### Added